New Folder and Attempted Practice Application
A new folder has been added to allow me to practice the skills I am learning in JS and put them to the test using a basic real world example.

I have also had an attempt at creating a 'Car Inverntory' system that allows a user to create a single entry (for now) and then search for that entry. Based on the search criteria, the user will be prompted with the car details or an error message. Learning outcome - I need to study up on constructors and onject creation in JaveScript.